Rename pkg/hf to pkg/hfmeta
pkg/hf (HuggingFace README/config metadata parsing) was easy to confuse with
pkg/backend/hf (the HTTP API handlers). Rename the metadata package to
hfmeta.
